Moved calculation of TransactionMass out of TransactionValidator, so t that it can be used in kaspawallet
