+++ domain/consensus/processes/blockvalidator/blockvalidator.go
@@ -4,6 +4,8 @@ import (
 	"math/big"
 	"time"
 
+	"github.com/kaspanet/kaspad/util/txmass"
+
 	"github.com/kaspanet/kaspad/domain/consensus/model"
 	"github.com/kaspanet/kaspad/domain/consensus/model/externalapi"
 	"github.com/kaspanet/kaspad/util/difficulty"
@@ -47,6 +49,8 @@ type blockValidator struct {
 	reachabilityStore   model.ReachabilityDataStore
 	consensusStateStore model.ConsensusStateStore
 	daaBlocksStore      model.DAABlocksStore
+
+	txMassCalculator *txmass.Calculator
 }
 
 // New instantiates a new BlockValidator
@@ -85,6 +89,8 @@ func New(powMax *big.Int,
 	reachabilityStore model.ReachabilityDataStore,
 	consensusStateStore model.ConsensusStateStore,
 	daaBlocksStore model.DAABlocksStore,
+
+	txMassCalculator *txmass.Calculator,
 ) model.BlockValidator {
 
 	return &blockValidator{
@@ -123,5 +129,7 @@ func New(powMax *big.Int,
 		reachabilityStore:   reachabilityStore,
 		consensusStateStore: consensusStateStore,
 		daaBlocksStore:      daaBlocksStore,
+
+		txMassCalculator: txMassCalculator,
 	}
 }
