+++ domain/consensus/processes/transactionvalidator/transactionvalidator.go
@@ -3,6 +3,7 @@ package transactionvalidator
 import (
 	"github.com/kaspanet/kaspad/domain/consensus/model"
 	"github.com/kaspanet/kaspad/domain/consensus/utils/txscript"
+	"github.com/kaspanet/kaspad/util/txmass"
 )
 
 const sigCacheSize = 10_000
@@ -16,32 +17,25 @@ type transactionValidator struct {
 	ghostdagDataStore          model.GHOSTDAGDataStore
 	daaBlocksStore             model.DAABlocksStore
 	enableNonNativeSubnetworks bool
-	massPerTxByte              uint64
-	massPerScriptPubKeyByte    uint64
-	massPerSigOp               uint64
 	maxCoinbasePayloadLength   uint64
 	sigCache                   *txscript.SigCache
 	sigCacheECDSA              *txscript.SigCacheECDSA
+	txMassCalculator           *txmass.Calculator
 }
 
 // New instantiates a new TransactionValidator
 func New(blockCoinbaseMaturity uint64,
 	enableNonNativeSubnetworks bool,
-	massPerTxByte uint64,
-	massPerScriptPubKeyByte uint64,
-	massPerSigOp uint64,
 	maxCoinbasePayloadLength uint64,
 	databaseContext model.DBReader,
 	pastMedianTimeManager model.PastMedianTimeManager,
 	ghostdagDataStore model.GHOSTDAGDataStore,
-	daaBlocksStore model.DAABlocksStore) model.TransactionValidator {
+	daaBlocksStore model.DAABlocksStore,
+	txMassCalculator *txmass.Calculator) model.TransactionValidator {
 
 	return &transactionValidator{
 		blockCoinbaseMaturity:      blockCoinbaseMaturity,
 		enableNonNativeSubnetworks: enableNonNativeSubnetworks,
-		massPerTxByte:              massPerTxByte,
-		massPerScriptPubKeyByte:    massPerScriptPubKeyByte,
-		massPerSigOp:               massPerSigOp,
 		maxCoinbasePayloadLength:   maxCoinbasePayloadLength,
 		databaseContext:            databaseContext,
 		pastMedianTimeManager:      pastMedianTimeManager,
@@ -49,5 +43,6 @@ func New(blockCoinbaseMaturity uint64,
 		daaBlocksStore:             daaBlocksStore,
 		sigCache:                   txscript.NewSigCache(sigCacheSize),
 		sigCacheECDSA:              txscript.NewSigCacheECDSA(sigCacheSize),
+		txMassCalculator:           txMassCalculator,
 	}
 }
